+++ src/extension.ts
@@ -12,7 +12,7 @@ export function activate(context: vscode.ExtensionContext) {
 
     // Use the console to output diagnostic information (console.log) and errors (console.error)
     // This line of code will only be executed once when your extension is activated
-    console.log('Congratulations, your extension "keithstudy" is now active!');
+    console.log('Extionsion is active, please enjoy it :)');
 
     let definitionDispose = vscode.languages.registerDefinitionProvider(
         ROBOT_MODE, new RobotDefinitionProvider());