+++ src/robotModels/TestSuite.ts
@@ -58,7 +58,7 @@ export class TestSuite {
     /**
      * give specific line in the document, locate the actual test case in the suite object
      * @param cursorLine the line of document we used for locate test case index
-     * @return The TestCase object to sea
+     * @return The TestCase object to search, and if the cursor is not in a test case, then return null
      */
     public locateToTestCase(cursorLine : number) : TestCase {
         // no test case in the current test suite, just return null